+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Nameplate Health
 
+## [v1.0.2](https://github.com/rbgdevx/nameplate-health/releases/tag/v1.0.2) (2024-12-03)
+
+- updating nameplate anchor code
+- minor variable cleanup
+
 ## [v1.0.1](https://github.com/rbgdevx/nameplate-health/releases/tag/v1.0.1) (2024-11-27)
 
 - adding wow interface id